requirements.txt
Django==1.7.5
argparse==1.2.1
beautifulsoup4==4.3.2
djangorestframework==3.1.0
psycopg2==2.6
requests==2.5.3
wsgiref==0.1.2
#django-wham  #sorry, I haven't uploaded the latest version to PyPI yet. It's up at https://testpypi.python.org/pypi/django-wham/0.1.3
